feat, docs(linear_algebra_number): Partial_Eq<LinAlgNumber> rewrote the first method to always return false if the variants mismatch. Also wrote docs for that function. All tests compiles.
At first I implemented the logic using casting. But since this will be used in a scientific environment, I would surmise that precision has a meaning and that no matter how small the difference between 3.0000001 and 3.0000000000000001 might be, that it's significant enough to be accounted for. As a result, only types with same variant can ever be truly equal.